Revert "Add Windows support for the Python API (#1811)"
This reverts commit ff491abf1f00238ea775c6076a8a12e94f005fd0.

+++ apis/python/src/tiledbsoma/io/ingest.py
@@ -1163,7 +1163,7 @@ def _write_dataframe(
         else:
             df.rename(columns={"index": id_column_name}, inplace=True)
 
-    df[SOMA_JOINID] = np.asarray(axis_mapping.data, dtype=np.int64)
+    df[SOMA_JOINID] = np.asarray(axis_mapping.data)
 
     df.set_index(SOMA_JOINID, inplace=True)
